Fix login validation and clean up routes
