--- .github/workflows/pylint.yml
@@ -1,23 +1,0 @@
-name: Pylint
-
-on: [push]
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ["3.8", "3.9", "3.10"]
-    steps:
-      - uses: actions/checkout@v4
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v3
-        with:
-          python-version: ${{ matrix.python-version }}
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install pylint fastapi pyrebase4 firebase-admin pydantic requests uvicorn
-      - name: Analyzing the code with pylint
-        run: |
-          pylint $(git ls-files '*.py')
