+++ app/models.py
@@ -36,22 +36,23 @@ class LoginSchema(BaseModel):
     """
     email: str
     password: str
-    
-def validate_user_data(user_data: BaseModel):
-    pattern = r'^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$'
 
-    if not re.match(pattern, user_data.email):
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Invalid email format"
-        )
+    @staticmethod
+    def validate_user_data(user_data: BaseModel):
+        pattern = r'^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$'
 
-    if len(user_data.password) < 7:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Password must be at least 8 characters long"
-        )
-    return user_data
+        if not re.match(pattern, user_data.email):
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail="Invalid email format"
+            )
+
+        if len(user_data.password) < 7:
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail="Password must be at least 8 characters long"
+            )
+        return user_data
 
 class ProgressStep(BaseModel):
     done: bool
